ai-engine/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
spacy==3.8.14
sentence-transformers==3.4.1
scikit-learn==1.6.1
transformers==4.51.3
torch==2.6.0
numpy==2.2.4
accelerate==1.4.0
peft==0.14.0
bitsandbytes==0.45.0
datasets==3.2.0
wandb==0.19.1
python-dotenv==1.0.1
requests==2.32.3
